[build] Also default to use a sysroot when the HOST_OS is linux
Previously, gn.py would not enable the sysroot when building for android
on Linux. However, the build also uses the host toolchain to create
tools and these need to be built with the sysroot. Otherwise, the new
goma backend refuses to run the build actions.

b/144018396

Change-Id: Ifda03ac82ce7d4cc8289198924a8316ab6751f45
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/151389
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Commit-Queue: Alexander Thomas <athom@google.com>