New 'non-type member completions' relevance tests in the analysis server.
This test includes: a failing type parameter test, as well as a common completion bug with Flutter widget construction, this test shows that the new relavance logic does the right thing.

When we add additional categories in completion metrics to split apart this group of completions, this test may be split as well, this name seemed appropriate for the time being.

Change-Id: I67638a121528be5e9a20bebc022a576282db1867
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/151631
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Jaime Wren <jwren@google.com>