Consolidate to a single implementation of refineBinaryExpressionType.
Previously, we had one implementation in TypeSystemImpl that did some
of the work, and then deferred to the superclass implementation in
TypeSystem to do the rest of the work.  But the separation between the
two didn't add any value, since all instances of TypeSystem are also
instances of TypeSystemImpl.  So we move all the logic into
TypeSystemImpl for clarity.

Change-Id: I2901ab854a1507d65545ab05e3662913f219ad56
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/152704
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>