Integrate suggested changes from the PR to publish the docs
Please look this over carefully; all of your changes have been integrated,
but not all of the changes are yours. While itegrating your suggestions I
found a few more occurrances of 'is not' and 'are not' that I converted.
I also found a couple of other things that needed to be updated.

Change-Id: Iddaaa22d7e584506bc3c16606166771fd5a75caf
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/152942
Commit-Queue: Kathy Walrath <kathyw@google.com>
Reviewed-by: Kathy Walrath <kathyw@google.com>