[vm/compiler] Make precompiler prune invocation dispatchers the same way we do for regular/closure functions
The precompiler has a notion of when a function should get its code
generated, when it should retain a functions types. This information
is used to prune function objects after AOT compilation.

This CL introduces the same pruning logic which we do for regular
functions as well as closure functions to invocation dispatchers.

Change-Id: I5c8a75f0cc822c71e876b8e2c9648d9976745895
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/154002
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Tess Strickland <sstrickl@google.com>