Fix the IANA link from Encoding.getByName docs
The previous version include the closing parenthesis in the link on the
dart doc site so it won't lead to the right place when clicked.

Switch to a markdown link instead of a raw URL to avoid ambiguity and
improve readability of the rendered doc.

Do some minor rephrasing of the rest of the comment to remove noise
words, format according to effective Dart.

Change-Id: I497cfeeeff0885eb34e798c469f95dc46f4a81f9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/156342
Auto-Submit: Nate Bosch <nbosch@google.com>
Commit-Queue: Nate Bosch <nbosch@google.com>
Commit-Queue: Kevin Moore <kevmoo@google.com>
Reviewed-by: Kevin Moore <kevmoo@google.com>